project/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-learn>=0.24.0
opencv-python>=4.5.0
networkx>=2.6.0
rdkit>=2021.09.0
tqdm>=4.62.0
pandas>=1.3.0
seaborn>=0.11.0
wandb>=0.12.0
tensorboard>=2.7.0
requests>=2.25.0
pyyaml>=5.4.0
jupyter>=1.0.0
ipython>=7.27.0